Fixed Patreon Link (Needs Review)
The link was previously going to miaberth which gives a 404 on Patreon. 

Searching for berthrage in Patreon gave me this link.
https://www.patreon.com/berthrage

The profile image matches you're GithHub Profile.

That being said I don't know if it's impersonation so please verify the changes and only proceed with merging if it looks good.